+++ src/Hash.h
@@ -54,13 +54,13 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include "LVB.h"
 
-long TopologicalHashComparison(Dataptr restrict, unsigned long, const TREESTACK_TREE_BRANCH *const, Lvb_bool, std::string, unsigned long&, const TREESTACK_TREE_BRANCH *const, TREESTACK *, long,
+long TopologicalHashComparison(Dataptr restrict, unsigned long long, const TREESTACK_TREE_BRANCH *const, Lvb_bool, std::string, unsigned long long&, const TREESTACK_TREE_BRANCH *const, TREESTACK *, long,
                                 const TREESTACK_TREE_BRANCH *const, const long, Lvb_bool);
-long HashComparison(unsigned long, unsigned long, Dataptr restrict, const TREESTACK_TREE_BRANCH *const, TREESTACK *, Lvb_bool, long, const TREESTACK_TREE_BRANCH *const, const long, Lvb_bool);
+long HashComparison(unsigned long long, unsigned long long, Dataptr restrict, const TREESTACK_TREE_BRANCH *const, TREESTACK *, Lvb_bool, long, const TREESTACK_TREE_BRANCH *const, const long, Lvb_bool);
 long CompareHashTreeToHashstack(Dataptr, TREESTACK *, const TREESTACK_TREE_BRANCH *const, const long, Lvb_bool);
 std::string ConvertSiteSetToString(Dataptr MSA, Objset *oset_1);
 std::string MakeHashSet(Dataptr restrict, const TREESTACK_TREE_BRANCH *const tree_2, const long);
-unsigned long HashSiteSet(std::string);
+unsigned long long HashSiteSet(std::string);
 long CollisionResolution(Dataptr restrict, TREESTACK *, const TREESTACK_TREE_BRANCH *const, Lvb_bool);
 
 #endif  // SRC_HASH_H_
